Переносит функцию для показа ошибки загрузки фотографий в модуль requests.
